requirements.txt
redis
openai
anthropic
httpx
python-dotenv
